+++ src/test/java/nextstep/ladder/domain/LadderTest.java
@@ -19,7 +19,7 @@ public class LadderTest {
         NonConsecutiveFlagGeneratorFake fakeGenerator = createFakeGenerator(booleans);
         Ladder ladder = LadderFactory.ladder(2, 2, fakeGenerator);
         assertThat(ladder).isEqualTo(
-                new Ladder(2, List.of(
+                new LadderImpl(2, List.of(
                         LinesFactory.lines(booleans),
                         LinesFactory.lines(booleans))));
     }
@@ -31,14 +31,14 @@ public class LadderTest {
         NonConsecutiveFlagGeneratorFake fakeGenerator = createFakeGenerator(booleans);
         Ladder ladder = LadderFactory.ladder(2, 2, fakeGenerator);
         assertThat(ladder).isEqualTo(
-                new Ladder(2, List.of(
+                new LadderImpl(2, List.of(
                         LinesFactory.lines(booleans),
                         LinesFactory.lines(booleans))));
     }
 
     @Test
     public void 라인이_존재하지_않을_수_없다() {
-        assertThatIllegalArgumentException().isThrownBy(() -> new Ladder(2, Collections.emptyList()));
+        assertThatIllegalArgumentException().isThrownBy(() -> new LadderImpl(2, Collections.emptyList()));
     }
 
     @Test
